rework pool reg testing
