+++ test/test_utils.ts
@@ -85,7 +85,7 @@ export const Networks = {
 
 type TxHash = FixlenHexString<32>
 
-export function hashTxBody(txBodyHex: string): TxHash {
+function hashTxBody(txBodyHex: string): TxHash {
     let b2 = blake2.createHash("blake2b", { digestLength: 32 })
     b2.update(Buffer.from(txBodyHex, 'hex'))
     return parseModule.parseHexStringOfLength(b2.digest('hex'), 32, InvalidDataReason.INVALID_B2_HASH)
@@ -100,6 +100,8 @@ export type NetworkIdlessTestResult = {
 export function bech32_to_hex(str: string): string {
     return utils.buf_to_hex(utils.bech32_decodeAddress(str))
 }
+
+export const DontRunOnLedger: string = "DO NOT RUN ON LEDGER"
 
 export function describeRejects(name: string, testList: any) {
     describe(name + "_JS", async () => {
@@ -141,6 +143,9 @@ export function describeRejects(name: string, testList: any) {
     
         for (const {testname, tx, additionalWitnessPaths, signingMode, errCls, errMsg } of testList) {
             it(testname, async() => {
+                if (errMsg === DontRunOnLedger) {
+                    return
+                }
                 const response = ada.signTransaction({
                     tx,
                     signingMode,
@@ -151,3 +156,40 @@ export function describeRejects(name: string, testList: any) {
         }
     })
 }
+
+export function describePositiveTest(name: string, tests: any[]) {
+    describe(name, async () => {
+        let ada: Ada = {} as Ada
+
+        beforeEach(async () => {
+            ada = await getAda()
+        })
+    
+        afterEach(async () => {
+            await (ada as any).t.close()
+        })
+    
+        for (const { testname, tx, signingMode, additionalWitnessPaths, txBody, result: expected } of tests) {
+            const additionalWitnessPathsIfPresent = additionalWitnessPaths || []
+            it(testname, async () => {
+                if (!txBody) {
+                    console.log("No tx body given!")
+                } else if (hashTxBody(txBody) !== expected.txHashHex) {
+                    console.log("Tx body hash mismatch")
+                }
+                const response = await ada.signTransaction({
+                    tx,
+                    signingMode,
+                    additionalWitnessPaths: additionalWitnessPathsIfPresent,
+                })
+                const networklessResponse: NetworkIdlessTestResult = {
+                    txHashHex: response.txHashHex,
+                    witnesses: response.witnesses,
+                    auxiliaryDataSupplement: response.auxiliaryDataSupplement,
+                }
+                expect(networklessResponse).to.deep.equal(expected)
+            })
+        }
+    })
+}
+
